refactor(iocuak-core): extract functions from isMetadataArrayLoadable
