+++ packages/iocuak-core/src/containerModuleMetadata/actions/decrementMetadataDependency.ts
@@ -1,0 +1,16 @@
+import { ContainerModuleMetadata } from '../models/ContainerModuleMetadata';
+
+export function decrementMetadataDependency(
+  metadata: ContainerModuleMetadata,
+  metadataToDependenciesMap: Map<ContainerModuleMetadata, number>,
+): number {
+  let metadataDependencies: number = metadataToDependenciesMap.get(
+    metadata,
+  ) as number;
+
+  metadataDependencies -= 1;
+
+  metadataToDependenciesMap.set(metadata, metadataDependencies);
+
+  return metadataDependencies;
+}
